+++ src/Wallet/WalletUtils.h
@@ -19,6 +19,7 @@
 
 #include <string>
 #include <cinttypes>
+#include <vector>
 
 #include "IWallet.h"
 #include "CryptoNoteCore/Currency.h"
@@ -26,7 +27,7 @@
 
 namespace CryptoNote {
 
-uint16_t getDefaultMixinByHeight(const uint64_t height);
+uint16_t getDefaultMixinByHeight(const uint32_t height);
 void throwIfKeysMismatch(const Crypto::SecretKey& secretKey, const Crypto::PublicKey& expectedPublicKey,
                          const std::string& message = "");
 bool validateAddress(const std::string& address, const CryptoNote::Currency& currency);
